fix nav event
